extract: Support patchelf
As version 0.18 fixed the long standing problems that tegra has had, we
can finally switch to that as opposed to other hacks. Unfortunately, as
older versions are known broken, we need to enforce a specific version.
This uses the same variable format as Lineage common extract and will
first check for the lineage prebuilt. If that's not found, it will try
to use a host copy, but only if the version matches what's requested.

Change-Id: Ia182eed374ec718113e86c188515d5c3882baffe